refactor: extract the ruleset public-page read model
convex/lib/rulesetDetailPage.ts owns slug resolution, the single
soft-delete gate (previously written three times), the faction listing
with its two documented degradation rules (null identity for
unparseable data; id fallback for a missing name), and viewer access;
both page queries become thin wrappers over it. Wire shapes unchanged,
pinned by the existing validators and suites.

Closes the ruleset item from the cluster-2 review.
